After doing optimizations in the parallel for body. Extra speed up achieved by removing the .size() on vector

+++ GameOfLife/Settings.h
@@ -12,7 +12,7 @@ static const int UNIVERSE_Y_COUNT = 10;
 static const int DEFAULT_TOTAL_TIME_STEPS = 10;
 
 static const int DEFAULT_INIT_LIVE_CELLS = 10;
-static const double DEFAULT_INIT_LIVE_CELLS_PROPORTION = 0.8;
+static const float DEFAULT_INIT_LIVE_CELLS_PROPORTION = 0.8f;
 static const int DEFAULT_NEIGHBORHOOD_SIZE = 9;
 
 //Any live cell with fewer than two live neighbours dies, as if caused by under - population.
